feat: Added render test
