+++ package.json
@@ -26,13 +26,17 @@
     "@babel/preset-typescript": "7.7.2",
     "@react-native-community/eslint-config": "0.0.5",
     "@storybook/react": "5.2.6",
+    "@types/enzyme": "3.10.3",
     "@types/faker": "4.1.7",
     "@types/jest": "24.0.23",
     "babel-loader": "8.0.6",
     "babel-plugin-lodash": "3.3.4",
+    "enzyme": "3.10.0",
+    "enzyme-adapter-react-16": "1.15.1",
     "faker": "4.1.0",
     "husky": "3.0.9",
     "jest": "24.9.0",
+    "jest-enzyme": "7.1.2",
     "lerna": "3.18.4",
     "lint-staged": "9.4.2",
     "npm-run-all": "4.1.5",
