+++ packages/templates/package.json
@@ -7,5 +7,15 @@
     "babel-plugin-cssta": "0.9.0",
     "babel-plugin-macros": "2.6.1",
     "cssta": "0.9.4"
+  },
+  "jest": {
+    "setupFilesAfterEnv": ["<rootDir>/setup.enzyme.js"],
+    "transform": {
+      "^.+\\.(t|j)sx?$": "babel-jest"
+    },
+    "collectCoverage": true
+  },
+  "scripts": {
+    "test": "jest"
   }
 }
