+++ packages/templates/setup.enzyme.js
@@ -1,0 +1,4 @@
+import 'jest-enzyme'
+import Enzyme from 'enzyme'
+import Adapter from 'enzyme-adapter-react-16'
+Enzyme.configure({ adapter: new Adapter() })
