+++ packages/templates/src/Layout.test.tsx
@@ -1,0 +1,9 @@
+import { shallow } from 'enzyme'
+import { Layout } from './Layout'
+
+describe('Layout', () => {
+  it('render', () => {
+    const wrapper = shallow(<Layout>hello</Layout>)
+    expect(wrapper).toMatchSnapshot()
+  })
+})
