+++ packages/templates/src/index.ts
@@ -1,1 +1,1 @@
-export * from './Formik'
+export * from './Layout'
