+++ packages/templates/src/story.tsx
@@ -5,19 +5,110 @@ import {
   PricingCard,
   ListItem,
   ThemeProvider,
+  Input as ElementsInput,
+  Button as ElementsButton,
 } from 'react-native-elements'
+import { Alert, Button, Text, TextInput, View } from 'react-native'
 import { random } from 'faker'
-import { View as Layout } from 'react-native'
-import cssta from 'cssta/native.macro'
-import { ElementsFormik, MyReactNativeForm } from './Formik'
+import { string, object } from 'yup'
+import { Formik, FormikProps } from 'formik'
+import { Layout } from './Layout'
 
-const StyledLayout = cssta(Layout)`
-  padding: 0px;
+const validationSchema = object().shape({
+  email: string().email(),
+})
 
-  @media (min-width: 600px) {
-    padding: 20px;
+interface FormValues {
+  email: string
+}
+
+const MyReactNativeForm = (_: FormikProps<FormValues>) => {
+  const initialValues: FormValues = {
+    email: '',
   }
-`
+  return (
+    <Formik
+      initialValues={initialValues}
+      onSubmit={values => {
+        Alert.alert(JSON.stringify(values))
+        console.log(values)
+      }}
+      validationSchema={validationSchema}
+    >
+      {({
+        handleChange,
+        handleBlur,
+        handleSubmit,
+        values,
+        errors,
+        isValid,
+        isSubmitting,
+        touched,
+      }) => (
+        <View>
+          <TextInput
+            placeholder="Input email address"
+            onChangeText={handleChange('email')}
+            onBlur={handleBlur('email')}
+            value={values.email}
+          />
+          <Button
+            onPress={handleSubmit}
+            title="Submit"
+            disabled={!values.email || !isValid || isSubmitting}
+          />
+          {touched.email && <Text>{errors.email}</Text>}
+        </View>
+      )}
+    </Formik>
+  )
+}
+
+const ElementsFormik = (_: FormikProps<FormValues>) => {
+  const initialValues: FormValues = {
+    email: '',
+  }
+  return (
+    <Formik
+      initialValues={initialValues}
+      onSubmit={(values, helpers) => {
+        Alert.alert(JSON.stringify(values))
+        console.log(values)
+        helpers.resetForm({ values: { email: '' } })
+      }}
+      validationSchema={validationSchema}
+    >
+      {({
+        handleChange,
+        handleBlur,
+        handleSubmit,
+        values,
+        errors,
+        isValid,
+        isSubmitting,
+        touched,
+      }) => (
+        <View>
+          <ElementsInput
+            label="email"
+            placeholder="Input email address"
+            onChangeText={handleChange('email')}
+            onBlur={handleBlur('email')}
+            value={values.email}
+            errorMessage={
+              touched.email && errors.email ? errors.email : undefined
+            }
+          />
+          <ElementsButton
+            onPress={handleSubmit}
+            title="Submit"
+            disabled={!values.email || !isValid || isSubmitting}
+          />
+        </View>
+      )}
+    </Formik>
+  )
+}
 
 const list = [
   {
@@ -33,18 +124,18 @@ const list = [
 storiesOf('templates', module)
   .addDecorator(story => <ThemeProvider>{story()}</ThemeProvider>)
   .add('Layout', () => (
-    <StyledLayout>
+    <Layout>
       <Header
         leftComponent={{ icon: 'menu', color: '#fff' }}
         centerComponent={{ text: 'MY TITLE', style: { color: '#fff' } }}
         rightComponent={{ icon: 'home', color: '#fff' }}
       />
-      <Layout
+      <View
         style={{
           flexDirection: 'row',
         }}
       >
-        <Layout style={{ flex: 0.5 }}>
+        <View style={{ flex: 0.5 }}>
           <PricingCard
             color="#4f9deb"
             title="Free"
@@ -52,8 +143,8 @@ storiesOf('templates', module)
             info={['1 User', 'Basic Support', 'All Core Features']}
             button={{ title: 'GET STARTED', icon: 'flight-takeoff' }}
           />
-        </Layout>
-        <Layout style={{ flex: 0.5 }}>
+        </View>
+        <View style={{ flex: 0.5 }}>
           <PricingCard
             color="#4f9deb"
             title="Free"
@@ -61,8 +152,8 @@ storiesOf('templates', module)
             info={['1 User', 'Basic Support', 'All Core Features']}
             button={{ title: 'GET STARTED', icon: 'flight-takeoff' }}
           />
-        </Layout>
-      </Layout>
+        </View>
+      </View>
       {list.map((item, i) => (
         <ListItem
           key={i}
@@ -72,7 +163,7 @@ storiesOf('templates', module)
           chevron
         />
       ))}
-    </StyledLayout>
+    </Layout>
   ))
   .add('MyReactNativeForm', () => <MyReactNativeForm />)
   .add('ElementsFormik', () => <ElementsFormik />)
